# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.PairingRequest:
    properties:
      session_id:
        type: string
      signature:
        type: string
    required:
    - session_id
    - signature
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Auth Service for B2B Suite – admin, employee, attendance, etc.
  title: B2B Suite API
  version: "1.0"
paths:
...
  /api/v1/web/login/cleanup:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: Cleanup result
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Cleanup failed
          schema:
            additionalProperties: true
            type: object
      summary: Cleanup expired sessions
      tags:
      - pairing
  /api/v1/web/login/confirm:
    post:
      consumes:
      - application/json
      parameters:
      - description: Session ID
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Token pair
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      summary: Confirm pairing
      tags:
      - pairing
  /api/v1/web/login/pair:
    post:
      consumes:
      - application/json
      description: Authenticates and pairs the device with a QR session.
      parameters:
      - description: Pairing request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.PairingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Pairing successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
      summary: Pair device
      tags:
      - pairing
  /api/v1/web/login/qr:
    get:
      description: Generates a new QR session for mobile-web pairing.
      produces:
      - application/json
      responses:
        "200":
          description: QR data and session ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Generation failed
          schema:
            additionalProperties: true
            type: object
      summary: Generate QR code
      tags:
      - pairing
  /api/v1/web/login/status:
    get:
      parameters:
      - description: Session ID
        in: query
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Status details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing session_id
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Session not found
          schema:
            additionalProperties: true
            type: object
      summary: Get pairing status
      tags:
      - pairing
  /api/v1/web/login/ws:
    get:
      description: Upgrades to WebSocket for real-time session communication.
      parameters:
      - description: Session ID to associate with the connection
        in: query
        name: session_id
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
        "400":
          description: Missing session_id
          schema:
            additionalProperties: true
            type: object
      summary: WebSocket connection endpoint
      tags:
      - websocket
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"